requirements.txt
azure-common==1.1.28
azure-core==1.29.4
azure-loganalytics==0.1.1
certifi==2023.7.22
charset-normalizer==3.3.0
greenlet==2.0.2
idna==3.4
importlib-metadata==6.7.0
isodate==0.6.1
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.3
msrest==0.7.1
oauthlib==3.2.2
pyodbc==4.0.39
requests==2.31.0
requests-oauthlib==1.3.1
six==1.16.0
SQLAlchemy==2.0.22
typing_extensions==4.7.1
urllib3==2.0.7
zipp==3.15.0
